src/test/java/lambda/DiapasonFunctionTest.java
package lambda;

import org.junit.Test;

import java.util.Arrays;
import java.util.List;

import static org.hamcrest.Matchers.is;
import static org.junit.Assert.assertThat;

public class DiapasonFunctionTest {

    @Test
    public void whenLinearFunctionThenLinearResults() {
        List<Double> result = DiapasonFunction.diapason(5, 8, x -> 2 * x + 1);
        List<Double> expected = Arrays.asList(11D, 13D, 15D);
        assertThat(result, is(expected));
    }

    @Test
    public void whenQuadraticFunction() {
        List<Double> result = DiapasonFunction.diapason(1, 4, x -> x * x + 2 * x + 1);
        List<Double> expected = Arrays.asList(4D, 9D, 16D);
        assertThat(result, is(expected));
    }

    @Test
    public void whenExponentialFunction() {
        List<Double> result = DiapasonFunction.diapason(-1, 2, x -> Math.pow(2, x));
        List<Double> expected = Arrays.asList(0.5D, 1D, 2D);
        assertThat(result, is(expected));
    }
}